Redis CommandCacheStorage
