+++ Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-teloxide = { version = "0.12", default-features = false, features = ["macros", "webhooks-axum", "rustls", "ctrlc_handler"] }
+teloxide = { version = "0.12", default-features = false, features = ["macros", "webhooks-axum", "rustls", "ctrlc_handler", "redis-storage"] }
 log = "0.4"
 pretty_env_logger = "0.5.0"
 tokio = { version =  "1.35.1", default-features = false, features = ["rt-multi-thread", "macros"] }
